// frontend/src/modules/product/types.ts
export enum EProductActionTypes {
	ADD_COMMENT = 'ADD_COMMENT',
	REMOVE_COMMENT = 'REMOVE_COMMENT',
	SET_PRODUCT_DATA = 'SET_PRODUCT_DATA',
	SET_PRODUCT_ERROR = 'SET_PRODUCT_ERROR',
	SET_PRODUCT_LOADING = 'SET_PRODUCT_LOADING',
	SET_PRODUCT_CATEGORIES = 'SET_PRODUCT_CATEGORIES',
	SET_PRODUCTS = 'SET_PRODUCTS',
	SET_LAST_PAGE = 'SET_LAST_PAGE',
}